UrlShorter-main/requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
sqlalchemy==2.0.37
requests==2.32.3
pymysql==1.1.1
python-dotenv==1.0.1
jinja2==3.1.5